results logged with success

+++ app/src/main/java/com/vb/yelplite/app/data/RemoteBusiness.kt
@@ -1,10 +1,7 @@
 package com.vb.yelplite.app.data
 
-import retrofit2.Response
-import retrofit2.http.GET
-
 data class RemoteBusiness(
-    val rating: Int,
+    val rating: Float,
     val price: String,
     val phone: String,
     val id: String,
